Better support for the LabelByChildAttribute across Toolbox Editors
