+++ Assets/Editor Toolbox/Runtime/Attributes/ILabelProcessorAttribute.cs
@@ -1,0 +1,8 @@
+﻿namespace Toolbox.Attributes
+{
+    /// <summary>
+    /// Temporary interface used to wrap all attributes used to override property labels.
+    /// </summary>
+    public interface ILabelProcessorAttribute
+    { }
+}
