+++ Assets/Editor Toolbox/Runtime/Attributes/ILabelProcessorAttribute.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 10b9c25a77d5f3a4291c427bbfbc88f4
+guid: 8be4122d49ab8134e92c17b380601f76
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
