+++ Assets/Examples/Scenes/SampleScene.unity
@@ -703,28 +703,6 @@ MonoBehaviour:
   sceneReference: {fileID: 0}
   childReference: {fileID: 0}
   prefabReference: {fileID: 0}
-  sampleField:
-    var1: {fileID: 2100000, guid: 7404c70251f9d0045a4aabaa49d83963, type: 2}
-    var2: 0
-    var3:
-      var1: 0
-      var2: Label By Value
-  sampleFields:
-  - var1: {fileID: 0}
-    var2: 64
-    var3:
-      var1: 0
-      var2: 
-  - var1: {fileID: 0}
-    var2: 53
-    var3:
-      var1: 0
-      var2: 
-  - var1: {fileID: 0}
-    var2: 38
-    var3:
-      var1: 0
-      var2: 
   bigNumber: 12345678
   currency: 20.41
 --- !u!4 &959025299
@@ -1171,6 +1149,16 @@ MonoBehaviour:
   var1: 25.4
   var2: {fileID: 977748988}
   var3: 0
+  var4:
+  - var1: 36
+  - var1: 52
+  sampleField:
+    var1: {fileID: 0}
+    var2: 266
+    var3:
+      var1: 0
+      var2: New Label from var2.var3
+  sampleFields: []
 --- !u!4 &1583836799
 Transform:
   m_ObjectHideFlags: 2
