+++ Assets/Examples/Scripts/SampleBehaviour7.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 [ExecuteAlways]
 [AddComponentMenu("Editor Toolbox/Cheat Sheet 7 (Special & Others)")]
@@ -16,4 +17,26 @@ public class SampleBehaviour7 : MonoBehaviour
 
     [HideLabel]
     public int var3;
+
+    [Label("LabelByChild", skinStyle: SkinStyle.Box)]
+
+    [LabelByChild("var3.var2")]
+    public SampleClass1 sampleField;
+    [LabelByChild("var2"), ReorderableList]
+    public SampleClass1[] sampleFields;
+
+    [Serializable]
+    public class SampleClass1
+    {
+        public Material var1;
+        public KeyCode var2;
+        public SampleClass2 var3;
+    }
+
+    [Serializable]
+    public class SampleClass2
+    {
+        public int var1;
+        public string var2;
+    }
 }
